Allow billing email to be empty
Allows passing `null` for the billing email to clear it. This means the
email must either be a string that is a valid email address or a null
value.

Since it is possible to clear the email address for an order in a
similar way via the admin-ajax endpoints, this is meant to provide
consistency with the wp-admin experience.