Add ReturnTypeWillChange attribute to classes implementing ArrayAccess
PHP 7.x and 8.1 will throw a fatal error, thus causing the unit tests
suite to abruptly end, if a method from an implemented interface doesn't
fully match the signature as defined in the interface. In WooCommerce
that's the case for the classes that implement ArrayAccess: in
PHP 8.0+ these signatures have type hints and return types (as in
"offsetUnset( mixed $offset ):void" ) but the code in WooCommerce
doesn't. We can't just add the missing types to the signatures because
then the same error appears in PHP 7.x, in which there aren't
types in the method signatures of ArrayAccess.

The fix consists of adding the ReturnTypeWillChange attribute,
introduced in PHP 8.1, which instructs the runtime to not throw the
error. PHP 7.x doesn't support attributes, but in this case the
attribute declaration is seen as a comment.

See https://php.watch/versions/8.1/ReturnTypeWillChange