Added setViewport
This change is done to ensure that these tests pass.

Wonky Tests:
- 'can create simple virtual product and add it to the cart'
- 'can create simple non-virtual product and add it to the cart'

It is observed that these tests fail when the following criteria are met:
- Running in headless mode
- Viewport is not set to 960x700

It is observed to pass when:
- Running in non-headless
- Running as an individual test file

The most likely culprit causing this issue is the verifyAndPublish function. More specifically, the element with the selector '.updated.notice' does not show up when the failing criteria are met.