Change '__return_true' to true in apply_filters()
Presumably the original author of this code intended to use true here
instead of '__return_true'; '__return_true' is commonly used with
add_filter() but it does not really make sense to use it here.

This was not really a bug because '__return_true' is a "truthy" value,
so the logic was working as intended; I am changing it to true just to
improve code readability.