Create and fill the product attributes lookup table in clean installs.
The table creation and filling was happening in a data migration, but
data migrations only run when WooCommerce is upgraded, not on a
clean install. Also, the option for using the lookup table
wasn't being set after a clean install (when the table has been created
and is empty since there are no products yet).

As part of the fix the table is added to the list of WooCommerce
core tables, and thus it's covered by the "Verify base database tables"
tool.