Refactor to use search_sku parameter
Instead of overloading the default search parameter.

This actually added a bit of complexity because it introduced the
possibility that a request could have both search and search_sku
parameters, and they might not be the same value.

I decided to handle the two cases differently: if both search strings
match, the assumption is that we're looking for a string in either the
content *or* the SKU field. However, if the the search strings are
different, the assumption is that we're looking for products with a
particular string in the content *and* a separate particular string
in the SKU.

New unit tests for this are still pending.