Actually sync orders from posts to custom orders table when needed.
This commit connects the logic of the DataSyncrhonizer class with
the actual posts to COT synchronization implemented in the
WPPostToCOTMigrator class, so orders are actually syncrhonized
using scheduled actions.

Additionally, a synchronization process is triggered whenever an order
is created or modified, using the woocommerce_after_order_object_save
hook.