Products API: revert get_objects method to `protected`
In 029257b I inadvertently changed the visibility of the get_objects
method to public. There is no need for its visibility to differ from
the parent class, and it ended up causing fatal errors for other classes
that are extending WC_REST_Products_Controller. This simply changes
the visibility back to protected.