Fix 'Invalid payment method' error upon double click 'Delete' #30862 (#30884)
* Fix 'Invalid payment method' error upon double click 'Delete' #30862

An 'Invalid payment method' error shows when a user double clicks the
'Delete' button in the 'Add Payment method screen'. Every time the user clicks
the button, there is a GET request to delete the payment method. First request is
handled correctly, but subsequent requests throw an error, because the payment
method cannot be found since it was already deleted.

This adds an event handler to disable the button on the first
click.

* Add changelog entry