Fix installation error in Onboarding Wizard when plugin generates unexpected output (#35866)
* Update plugin activation check

This is refactored to test the plugin state instead of checking the API call result.

This is because WordPress can return an error despite the plugin activating correctly. The `unexpected_output` error, in one such case, when WordPress detects stray output upon activation. To workaround this, the plugin is checked for its active status directly.

* Add tests

* Add changelog

* Fix lint issues